Conflicts:
	contrib/git_id/git_id_vc80.sln
	contrib/git_id/git_id_vc80.vcproj
	src/game/AchievementMgr.cpp
	src/game/BattleGround.cpp
	src/game/BattleGroundMgr.cpp
	src/game/Player.cpp
	src/game/SpellAuras.cpp
	src/game/SpellAuras.h
	src/game/SpellEffects.cpp
	win/VC100/game.vcxproj.filters

+++ src/shared/Database/Database.cpp
@@ -52,7 +52,7 @@ SqlPreparedStatement * SqlConnection::GetStmt( int nIndex )
         return NULL;
 
     //resize stmt container
-    if(m_holder.size() <= nIndex)
+    if((int)m_holder.size() <= (int)nIndex)
         m_holder.resize(nIndex + 1, NULL);
 
     SqlPreparedStatement * pStmt = NULL;
